update archivedoppage for component changes
